requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
langchain
langchain-community
langchain-openai
sentence-transformers
chromadb
tiktoken